+++ tests/testthat/test_PipeOpTorchAdaptiveAvgPool.R
@@ -37,7 +37,7 @@ test_that("PipeOpTorchAdaptiveAvgPool3D works with a 1d output size", {
   po_test = po("nn_adaptive_avg_pool3d", output_size = 10)
   task = nano_imagenet()
   graph = po("torch_ingress_ltnsr") %>>%
-    po("nn_reshape", shape = c(NA, 3, 64, 8, 8)) %>>%
+    po("nn_reshape", shape = c(-1, 3, 64, 8, 8)) %>>%
     po_test
 
   expect_pipeop_torch(graph, "nn_adaptive_avg_pool3d", task)
@@ -47,7 +47,7 @@ test_that("PipeOpTorchAdaptiveAvgPool3D works with a 3d output size", {
   po_test = po("nn_adaptive_avg_pool3d", output_size = c(10, 11, 12))
   task = nano_imagenet()
   graph = po("torch_ingress_ltnsr") %>>%
-    po("nn_reshape", shape = c(NA, 3, 64, 8, 8)) %>>%
+    po("nn_reshape", shape = c(-1, 3, 64, 8, 8)) %>>%
     po_test
 
   expect_pipeop_torch(graph, "nn_adaptive_avg_pool3d", task)
@@ -58,15 +58,15 @@ test_that("PipeOpTorchAdaptiveAvgPool3D paramtest", {
   expect_paramtest(res)
 })
 
-sampler_adaptive_avg_pool = function(dim, batch = TRUE) {
+sampler_adaptive_avg_pool = function(dim) {
   list(
-    shape_in = sample(20:25, size = dim + 1 + as.integer(batch), replace = TRUE),
+    shape_in = sample(20:25, size = dim + 2L, replace = TRUE),
     conv_dim = dim,
     output_size = sample(c(1, dim), size = 1)
   )
 }
 
-test_that("adaptive_avg_output_shape works when there is a batch dimension", {
+test_that("adaptive_avg_output_shape works", {
   for (dim in 1:3) {
     testcase = sampler_adaptive_avg_pool(dim)
     mg = switch(dim,
@@ -81,17 +81,33 @@ test_that("adaptive_avg_output_shape works when there is a batch dimension", {
   }
 })
 
-test_that("adaptive_avg_output_shape works when there is no batch dimension", {
-  for (dim in 1:3) {
-    testcase = sampler_adaptive_avg_pool(dim, batch = FALSE)
-    mg = switch(dim,
-      nn_adaptive_avg_pool1d,
-      nn_adaptive_avg_pool2d,
-      nn_adaptive_avg_pool3d
-    )
-    m = do.call(mg, testcase[names(testcase) %in% formalArgs(mg)])
-    outshape = with_no_grad(m(do.call(torch::torch_randn, args = list(unname(testcase$shape_in)))))$shape
-    expect_warning(shape <<- do.call(adaptive_avg_output_shape, args = testcase), regexp = "batch dimension")
-    expect_true(all(outshape == shape))
+test_that("adaptive_avg_output_shape requires a batch dimension", {
+  expect_error(
+    adaptive_avg_output_shape(shape_in = c(3, 20, 20), conv_dim = 2, output_size = 5),
+    "length 4"
+  )
+})
+
+test_that("shape inference matches the operator", {
+  expect_shape_inference("nn_adaptive_avg_pool2d", list(output_size = c(2, 3)), c(2, 3, 16, 20))
+  expect_shape_inference("nn_adaptive_avg_pool1d", list(output_size = 4), c(2, 3, 17))
+  expect_shape_inference("nn_adaptive_avg_pool3d", list(output_size = c(2, 3, 4)), c(2, 3, 5, 7, 9))
+})
+
+test_that("an unknown input extent still gives a known output", {
+  # the output size is fixed by `output_size`, so rejecting an unknown input extent would throw
+  # away information
+  for (d in 1:2) {
+    obj = po(sprintf("nn_adaptive_avg_pool%id", d), output_size = rep(4L, d))
+    shape_in = as.integer(c(NA, 3L, rep(NA_integer_, d)))
+    expect_equal(obj$shapes_out(list(input = shape_in))[[1L]], c(NA, 3L, rep(4L, d)))
+  }
+})
+
+test_that("shape inference agrees with the module for random shapes and parameters", {
+  for (d in 1:3) {
+    expect_shape_inference(sprintf("nn_adaptive_avg_pool%id", d),
+      params = function() list(output_size = sample(1:4, d)),
+      generators = gen_shape(d + 2L))
   }
 })